[ALPS05014853] mtu3: update USB driver
[Detail]
Update to latest codes

MTK-Commit-Id: 00accd57bb44174f9a9d7c7f9b33dcecec6be8b1

CR-Id: ALPS05014853
Feature: [Module]USB 2.0

+++ drivers/usb/mtu3/mtu3_gadget_ep0.c
@@ -16,8 +16,6 @@
  *
  */
 
-#include <linux/usb/composite.h>
-
 #include "mtu3.h"
 
 /* ep0 is always mtu3->in_eps[0] */
@@ -572,7 +570,7 @@ static void ep0_tx_state(struct mtu3 *mtu)
 	struct usb_request *req;
 	u32 csr;
 	u8 *src;
-	u8 count;
+	u16 count;
 	u32 maxp;
 
 	dev_dbg(mtu->dev, "%s\n", __func__);
